Add files via upload
This electronic schematic without PCB show how to measure by means of a microcontroller from ATMEL - ATSAMD21G18A and AFE4490 and ADS1292 the SPO2 and ECG ploting via software the graphics. This schematic known that ATSAMD21G18A has a LDO but is added another LDO to give more stability to the input voltage linked to a power inductor and ultra low power LDO from TI TPS7A4901 with the help of another componente LP3878SD-ADJ linked to other powe inductor to control LEDS activity and a power supply byTPS717xx with the help of Raspberry Pi v.3 and the LCD Touch Screen 7" for Raspberry pi and a Case Smarth PI linked to a SPI monitor or a case linked to HDMI monitor. So is welcome the help of a calibration componentes to assign an equipment validation. Thank you very much